gui: created wizard module and moved each wizard page in a separate file
