+++ circleguard/wizard/pages/__init__.py
@@ -1,0 +1,4 @@
+from .intro import IntroPage  # noqa: F401
+from .apikey import ApiKeyPage  # noqa: F401
+from .tips import TipsPage  # noqa: F401
+from .conclusion import ConclusionPage  # noqa: F401
